Pass to convert CUDA/HIP printf() calls to OpenCL
SPIRV-LLVM translator generates a wrong pointer address space to printf
format string if it's not the correct one in the input. This pass
converts the format string address space to constant before we pass
the IR to SPIRV emission.

Now converts the AS, but doesn't yet fix the return value which is
the number of processed format strings in CUDA and strlen() in AMD's
HIP.