Add a work around for a kernel rename bug
Encountered a bug where kernels are renamed unintentionally by
level0. The kernel names are appended with a suffix like ".1". This
bug occurs when SPIR-V binaries have entry points and functions by the
same name. Such binaries are valid SPIR-V because names, assigned to
functions via OpName, are not semantically meaningful.