OpenCL: Fix duplicate device reference in a context
This fixes broken tests under samples/hipSymbol/.

The double addition of the CHIPDevice to a CHIPContext caused multiple
CHIPDeviceVar instances per global-scope device variable to be created
in a way that the host and kernels saw different device pointers. This
then caused the host and kernels not to see each other's writes to the
variables.