Workaround for #48, adjust hip_async_interop sample
This commit provides a workaround for issue #48 by filtering out elements
in the embedded SPIR-V modules that trigger a suspected bug in certain
versions of Intel Compute Runtime.

Adjusted the SPIR-V binary in the kernel.cpp for avoiding firing the
suspected bug in Intel Compute Runtimes. Few OpName and OpDecorate was
removed from the binary.

Fixes #48