Fix compiler error with __attribute__((__noinline__))
Fix a compiler error caused by C++ libraries (e.g. libstdc++ provided
by g++-12) using `__attribute__((__noinline__))` in their code. Our
definition of the `__noinline__` appears in the compilation as followed:

```
/usr/lib/gcc/x86_64-linux-gnu/12/../../../../include/c++/12/bits/shared_ptr_base.h:196:37: error: use of undeclared identifier 'noinline'; did you mean 'inline'?
      __attribute__((__attribute__((noinline))))
...
```

For Clang <15 this issue is fixed by substituting the `__noinline__` with
empty string and, therefore, rendering it ineffective, unfortunately.

Fortunately, In Clang >=15 the `__noinline__` is a builtin keyword
and the following use cases work in CUDA and HIP language modes:

```
__attribute__((__inline__)) void foo(...);
__inline__ void bar(...);
```