SPIR-V filter: Fix erased import linkage attribute
... produced invalid SPIR-V binary (function declarations must have
import linkage). Also, issue a warning when we encounter imported
functions - it's likely that the user has forgotten definitions for
device symbols.

This patch is known to fix a JIT issue which includes this compiler
error message: "Fails to verify module: Global is external, but
doesn't have external or weak linkage!"