Fix #211 and others (#395)
* AllocationTracker: Add allocationInfo visitor

* Add an adaptor for iterating only keys in maps

* Print pointer values in kernel launch debug messages

Print the value of the kernel pointer arguments in the kernel launch
debug messages.

* RegisterdVarCopy: sync all host and managed allocation

We need to do this as a baseline because some allocations might passed
to kernel indirectly (e.g. passed in aggregates or in other allocations)
and we don't have means to detect them (yet?).

* OpenCL: track SVM allocations with std::shared_ptr

This is needed for follow up commit, where we need to extend the
lifetimes of the SVM allocations despite being freed by an user
through hipFree().

* OpenCL&Level0: Annotate indirect buffer use

Inform driver about possible indirect buffer accesses so they get
synchronized properly.

* OpenCL: Unshare cl_kernel

Refactor the OpenCL backend (and the base backend as needed) so that
cl_kernel is not shared among threads (sharing cl_kernel is
discouraged by the OpenCL spec). This seems to fix nondeterministic
crashes in Unit_hipMemcpyWithStream_MultiThread test case.

* Remove unneeded "first touch" code

The code is not needed after properly informing the driver about
indirect buffer uses.

* Fix a test with false positive test outcome

The test did not return non-zero on error nor ctest was configured to
check output messages for test pass or failure.

* Add test for an indirect mapped host allocation

---------

Co-authored-by: Henry Linjamäki <henry.mikael.linjamaki@intel.com>