OpenCL: Implement CUDA callbacks using OCL CBs and user event hacks
hipStreamSemantics tests behavior that CUDA callbacks are asynchronously
executed in the background whereas OpenCL might or might not execute
them in the background. It could execute the callback in the user thread
in case the command it is calling back from was finished before the CB
was set.

Hack around this by creating a barrier which listens to a user event
which is triggered only after setting the CB, forcing parallel execution
of the callback. That is, in case the OpenCL platform supports parallel
execution of callbacks the first place. I don't believe that's required
to be the case, so this is down to implementation still.

Might Fix #421 and #418.