hipStreamSemantics: remove wrong hipStreamSynchronize(0)
hipStreamSynchronize on default stream doesn't finish only
the default stream, but also all the others on the same device:

  "Specifically, specifying the null stream will cause the command to
   wait for other streams on the same device to complete all pending
   operations."

https://rocm-developer-tools.github.io/HIP/group__Stream.html#gabbfb9f573a6ebe8c478605ecb5504a74

Thus, calling it here and expecting only the default stream's
work to get finished in parallel with the other stream's CB
is wrong. If it this test passed on an environment, I think
that's a bug in that environment since that synch should
also wait for the CB queue to finish.