OpenCL: Implement CUDA callbacks using OCL CBs and user event hacks (#422)
* OpenCL: Implement CUDA callbacks using OCL CBs and user event hacks

hipStreamSemantics tests behavior that CUDA callbacks are asynchronously
executed in the background whereas OpenCL might or might not execute
them in the background. It could execute the callback in the user thread
in case the command it is calling back from was finished before the CB
was set.

Hack around this by creating a barrier which listens to a user event
which is triggered only after setting the CB, forcing parallel execution
of the callback. That is, in case the OpenCL platform supports parallel
execution of callbacks the first place. I don't believe that's required
to be the case, so this is down to implementation still.

Might Fix #421 and #418.

* hipStreamSemantics: remove wrong hipStreamSynchronize(0)

hipStreamSynchronize on default stream doesn't finish only
the default stream, but also all the others on the same device:

  "Specifically, specifying the null stream will cause the command to
   wait for other streams on the same device to complete all pending
   operations."

https://rocm-developer-tools.github.io/HIP/group__Stream.html#gabbfb9f573a6ebe8c478605ecb5504a74

Thus, calling it here and expecting only the default stream's
work to get finished in parallel with the other stream's CB
is wrong. If it this test passed on an environment, I think
that's a bug in that environment since that synch should
also wait for the CB queue to finish.

* hipStreamSemantics: Catch 'Failed' strings for failure