OpenCL backend: add support for cl_intel_unified_shared_memory
Previously, the OpenCL device was checked for fine-grained SVM support,
and if it was unavailable, the backend would assume only coarse-grained
support and insert SVMMap & SVMUnmap for each buffer used by a kernel.

Since Intel OpenCL and PoCL both support the USM extension now, but
don't support fine-grained SVM, this commit helps to avoid the overhead
of the unnecessary Map & Unmap commands.