OpenCL BE: Do not pass -x spir
'-x spir' requests compilation of the old SPIR 1.2/2.0 whereas
we use the new SPIR-V which doesn't need the build flag, as
long as we call clCreateProgramWithIL(). This we do via the
C++ wrapper.

Passing -x spir might fail if the device doesn't support the old
SPIR even though it might support the new SPIR-V.