Remove a confusing already registered and mapped warning
Remove confusing "A device function is already registered and mapped
to a different module" warning. It just happens to be that there may
be multiple __hipRegisterFunction() calls to register same functions
due to the way the code is generated for the inline and template
functions. There is a bit more details in the code.