Adjustments for (upcoming) LLVM-23 and Clang's new offload driver
This patch updates chipStar to work on non-forked, upstream
LLVM-23. The recent Clang has switched to use its new offload driver
by default for HIP compilation and chipStar is adjusted to it. There
is a still an option to use the old while it still around in Clang.

That said, chipStar doesn't yet work on the recent Clang until the
following upstream patch lands:
https://github.com/llvm/llvm-project/pull/187655.